refactor: prune flask app and messagebus
