+++ runner/domain/events.py
@@ -7,12 +7,6 @@ class Event:
 
 
 @dataclass
-class InstructionsExecuted(Event):
-    job_id: str
-    instructions: List[Tuple[str, Dict]]
-
-
-@dataclass
 class ScriptExecuted(Event):
     jobid: str
     script: str
