+++ runner/tests/unit/test_services.py
@@ -1,53 +1,53 @@
-from typing import Tuple, List, Dict
-from runner.adapters import dispatcher
-from runner.service_layer import services
+import uuid
+from runner.service_layer import services, unit_of_work
 from runner.domain import model
 
 
-class FakeRepository(dict):
-    def list(self):
-        return self.values()
+class FakeJobUow(unit_of_work.AbstractJobUnitOfWork):
+    def __init__(self) -> None:
+        self.jobs = FakeJobRepository()
+        self.commited = False
+
+    def commit(self):
+        self.commited = True
 
 
-class FakeDispatcher(set):
-    def execute(self, instructions: List[Tuple[str, dict]], repo) -> str:
-        job_id = dispatcher.generate_job_id()
-        self.add(job_id)
-        return job_id
-
-    def has_executed(self, job_id: str) -> bool:
-        return job_id in self
+class FakeJobRepository(dict):
+    def add(self, job):
+        self[job.id] = job
 
 
-class FakeScript(model.AbstractScript):
-    required_args = tuple()
-
-    def __init__(self, ref: str) -> None:
-        self.ref = ref
-
-    def execute(self, args: Dict) -> None:
-        print(f"execute {self.ref}")
+class FakeExecuter:
+    def __call__(self, job):
+        pass
 
 
-def test_get():
-    script_ref = "fake_script"
-    repo = FakeRepository(fake_script=FakeScript(script_ref))
-
-    script_instance = services.get(script_ref, repo)
-
-    assert script_instance is not None
-    assert script_instance.ref == script_ref
+class ExecuteError(Exception):
+    pass
 
 
-def test_execute():
-    repo = FakeRepository(
-        open_file=FakeScript("open_file"),
-        update_file=FakeScript("update_file"),
-        release_assets=FakeScript("release_assets"),
-    )
-    dispatcher = FakeDispatcher()
-    instructions = [("open_file", {}), ("update_file", {}), ("release_assets", {})]
+class FailExecuter:
+    def __call__(self, job):
+        raise ExecuteError
 
-    job_id = services.execute(instructions, repo, dispatcher)
 
-    assert dispatcher.has_executed(job_id)
+def test_can_execute_job():
+    job = model.Job(uuid.uuid4(), instructions=[])
+    job_uow = FakeJobUow()
+    job_uow.jobs.add(job)
+
+    services.execute(job.id, job_uow, executer=FakeExecuter())
+
+    assert job.state == "completed"
+    assert job_uow.commited
+
+
+def test_fail_to_execute_job():
+    job = model.Job(uuid.uuid4(), instructions=[])
+    job_uow = FakeJobUow()
+    job_uow.jobs.add(job)
+
+    services.execute(job.id, job_uow, executer=FailExecuter())
+
+    assert job.state == "failed"
+    assert job_uow.commited
